+++ tests/testcases/ok_exports.pas
@@ -1,0 +1,24 @@
+{ Test parsing @code(exports) section. }
+unit ok_exports;
+
+interface
+
+uses
+  Windows;
+
+{ Description of Foo. }
+procedure Foo;
+
+{ Description of Bar. }
+procedure Bar(const S: TStr64);
+
+exports
+  Foo,
+  Bar;
+
+{ Description of Foo2. }
+procedure Foo2;
+
+implementation
+
+end.