+++ tests/testcases_output/php/ok_exports/docs.php
@@ -1,0 +1,8 @@
+<?php
+global $pasdoc;
+$pasdoc = array(
+  'ok_exports' => array('html_filename' => 'ok_exports.html', 'type' => 'unit'),
+  'Foo' => array('html_filename' => 'ok_exports.html#Foo', 'type' => 'procedure'),
+  'Bar' => array('html_filename' => 'ok_exports.html#Bar-TStr64-', 'type' => 'procedure'),
+  'Foo2' => array('html_filename' => 'ok_exports.html#Foo2', 'type' => 'procedure'),
+);